+++ src/Tests/Testing.Common/Testing/RefactoringVerifier.cs
@@ -52,6 +52,41 @@ public abstract class RefactoringVerifier<TRefactoringProvider> : CodeVerifier
             cancellationToken: cancellationToken);
     }
 
+    /// <summary>
+    /// Verifies that refactoring will be applied correctly using specified <typeparamref name="TRefactoringProvider"/>.
+    /// </summary>
+    /// <param name="file">Source file where text selection is marked with <c>[|</c> and <c>|]</c> tokens.</param>
+    public async Task VerifyRefactoringAsync(
+        TestFile file,
+        IEnumerable<AdditionalFile>? additionalFiles = null,
+        string? equivalenceKey = null,
+        TestOptions? options = null,
+        CancellationToken cancellationToken = default)
+    {
+        if (file is null)
+            throw new ArgumentNullException(nameof(file));
+
+        if (file.ExpectedSource is null)
+            throw new ArgumentException("Expected source is required.", nameof(file));
+
+        var code = TestCode.Parse(file.Source);
+        var expected = ExpectedTestState.Parse(file.ExpectedSource);
+
+        var data = new RefactoringTestData(
+            code.Value,
+            code.Spans.OrderByDescending(f => f.Start).ToImmutableArray(),
+            additionalFiles,
+            equivalenceKey: equivalenceKey,
+            directoryPath: file.DirectoryPath,
+            fileName: file.Name);
+
+        await VerifyRefactoringAsync(
+            data,
+            expected,
+            options,
+            cancellationToken: cancellationToken);
+    }
+
     internal async Task VerifyRefactoringAsync(
         string source,
         string sourceData,
@@ -106,7 +141,7 @@ public abstract class RefactoringVerifier<TRefactoringProvider> : CodeVerifier
 
             using (Workspace workspace = new AdhocWorkspace())
             {
-                (Document document, ImmutableArray<ExpectedDocument> expectedDocuments) = CreateDocument(workspace.CurrentSolution, data.Source, data.AdditionalFiles, options);
+                (Document document, ImmutableArray<ExpectedDocument> expectedDocuments) = CreateDocument(workspace.CurrentSolution, data.Source, directoryPath: data.DirectoryPath, fileName: data.FileName, data.AdditionalFiles, options);
 
                 SemanticModel semanticModel = (await document.GetSemanticModelAsync(cancellationToken))!;
 
@@ -204,6 +239,36 @@ public abstract class RefactoringVerifier<TRefactoringProvider> : CodeVerifier
     /// <summary>
     /// Verifies that refactoring will not be applied using specified <typeparamref name="TRefactoringProvider"/>.
     /// </summary>
+    /// <param name="file">Source file where text selection is marked with <c>[|</c> and <c>|]</c> tokens.</param>
+    public async Task VerifyNoRefactoringAsync(
+        TestFile file,
+        IEnumerable<AdditionalFile>? additionalFiles = null,
+        string? equivalenceKey = null,
+        TestOptions? options = null,
+        CancellationToken cancellationToken = default)
+    {
+        if (file is null)
+            throw new ArgumentNullException(nameof(file));
+
+        var code = TestCode.Parse(file.Source);
+
+        var data = new RefactoringTestData(
+            code.Value,
+            code.Spans,
+            additionalFiles: additionalFiles,
+            equivalenceKey: equivalenceKey,
+            directoryPath: file.DirectoryPath,
+            fileName: file.Name);
+
+        await VerifyNoRefactoringAsync(
+            data,
+            options,
+            cancellationToken: cancellationToken);
+    }
+
+    /// <summary>
+    /// Verifies that refactoring will not be applied using specified <typeparamref name="TRefactoringProvider"/>.
+    /// </summary>
     public async Task VerifyNoRefactoringAsync(
         RefactoringTestData data,
         TestOptions? options = null,
@@ -223,7 +288,7 @@ public abstract class RefactoringVerifier<TRefactoringProvider> : CodeVerifier
 
         using (Workspace workspace = new AdhocWorkspace())
         {
-            (Document document, ImmutableArray<ExpectedDocument> _) = CreateDocument(workspace.CurrentSolution, data.Source, data.AdditionalFiles, options);
+            (Document document, ImmutableArray<ExpectedDocument> _) = CreateDocument(workspace.CurrentSolution, data.Source, directoryPath: data.DirectoryPath, fileName: data.FileName, data.AdditionalFiles, options);
 
             SemanticModel semanticModel = (await document.GetSemanticModelAsync(cancellationToken))!;